requirements.txt
scrapy
networkx
requests
beautifulsoup4
matplotlib
scipy
python-whois
dnspython
geoip2
tldextract
selenium
webdriver-manager
dnspython
pydantic
pydantic_settings
